requirements.txt
anyio==3.5.0
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.75.2
h11==0.13.0
idna==3.3
peewee==3.14.10
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
